End day with passing tests
