+++ src/test/java/com/adobe/epubcheck/test/JsonCompare.java
@@ -1,4 +1,4 @@
-package org.idpf.epubcheck.common;
+package com.adobe.epubcheck.test;
 
 import java.io.File;
 import java.io.IOException;
