+++ src/test/java/com/adobe/epubcheck/test/OutputDifferenceListener.java
@@ -1,4 +1,4 @@
-package org.idpf.epubcheck.common;
+package com.adobe.epubcheck.test;
 
 import java.text.DateFormat;
 import java.text.ParseException;
