+++ src/test/java/com/adobe/epubcheck/test/OverriddenMessageDictionaryTest.java
@@ -9,7 +9,6 @@ import com.adobe.epubcheck.util.DefaultReportImpl;
 import java.io.File;
 import java.net.URL;
 
-import org.idpf.epubcheck.common.CommonTestRunner;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -20,7 +19,7 @@ public class OverriddenMessageDictionaryTest
   public void ensureOverridenMessages()
   {
     String testName = "severity_overrideOk.txt";
-    URL inputUrl = CommonTestRunner.class.getResource("command_line");
+    URL inputUrl = OverriddenMessageDictionaryTest.class.getResource("command_line");
     String inputPath = inputUrl.getPath();
     String overrideFile = inputPath + "/" + testName;
