+++ src/test/java/com/adobe/epubcheck/test/opf_Test.java
@@ -1,8 +1,6 @@
 package com.adobe.epubcheck.test;
 
-import org.idpf.epubcheck.common.CommonTestRunner;
-import org.idpf.epubcheck.common.CommonTestRunner.TestOutputType;
-import org.idpf.epubcheck.common.NoExitSecurityManager;
+import com.adobe.epubcheck.test.CommonTestRunner.TestOutputType;
 import org.junit.*;
