+++ src/test/java/org/idpf/epubcheck/api/ApiConstructorsTest.java
@@ -2,7 +2,7 @@ package org.idpf.epubcheck.api;
 
 import com.adobe.epubcheck.api.EpubCheck;
 import com.adobe.epubcheck.api.Report;
-import org.idpf.epubcheck.common.CommonTestRunner;
+import com.adobe.epubcheck.test.CommonTestRunner;
 import com.adobe.epubcheck.util.WriterReportImpl;
 import org.junit.Assert;
 import org.junit.Test;
