+++ src/test/resources/com/adobe/epubcheck/test/xhtml/dtd_expected_results.json
@@ -351,10 +351,10 @@
     "id" : "xhtmldtd",
     "fileName" : "OPS/xhtml1-strict.dtd",
     "media_type" : "application/xml-dtd",
-    "compressedSize" : 6760,
-    "uncompressedSize" : 25473,
+    "compressedSize" : 6768,
+    "uncompressedSize" : 25503,
     "compressionMethod" : "Deflated",
-    "checkSum" : "5887cf989685d6faf5ca2f1823df7c7040657ba11ff579af759de125143e2",
+    "checkSum" : "af6d0c95e945736f31ebe521bd024ccdb82ae12245d327b7f50c45323c7f5",
     "isSpineItem" : false,
     "spineIndex" : null,
     "isLinear" : false,
