+++ src/test/resources/org/idpf/epubcheck/pkg/cli/20-severity-tester/mimetype
@@ -1,0 +1,1 @@
+application/epub+zip
